All Requests Completed

+++ src/main/java/com/turkcell/rentACar/business/request/invoiceRequests/UpdateInvoiceRequest.java
@@ -23,5 +23,5 @@ public class UpdateInvoiceRequest {
     private double rentDay;
 
     @Positive
-    private int totalRentCarPrice;
+    private double totalRentCarPrice;
 }
